Changed sequence number to int

+++ src/edu/chl/asciicam/file/FileController.java
@@ -33,7 +33,7 @@ public class FileController {
 	
 	public static final String UNMOUNTED_SD = "SDCARD_NOT_MOUNTED";
 	
-	private static double SEQ_NUMBER = 0;
+	private static int SEQ_NUMBER = 0;
 	
 	/**
 	 * Just an empty constructor, this class does not save any output or inputstreams
@@ -92,7 +92,7 @@ public class FileController {
 	}
 
 	//Sequence for file storing
-	private double getSequenceNumber(){
+	private int getSequenceNumber(){
 		return SEQ_NUMBER;
 	}
 	//Increment sequence
@@ -133,7 +133,7 @@ public class FileController {
 	 * @param options A hashmap consisting of the options to be saved.
 	 */
 	public void saveOptions(HashMap<String, String> options){
-		
+		// TODO
 	}
 	
 	/**
@@ -141,6 +141,7 @@ public class FileController {
 	 * @return A map of all options saved.
 	 */
 	public HashMap<String, String> getOptions(){
+		// TODO
 		return new HashMap<String, String>();
 	}
